New Commit (#5)
* Added delete functionality to products page

* Created a productBatch template, and modified the productBatchController by assigning get and post mappings for "/productBatch". Product Batch template currently only has newProductBatch form. Added a button on portalLanding for the productBatch template.

* Created a productBatch template, and modified the productBatchController by assigning get and post mappings for "/productBatch". Product Batch template currently only has newProductBatch form. Added a button on portalLanding for the productBatch template.

* Created a productBatch template, and modified the productBatchController by assigning get and post mappings for "/productBatch". Product Batch template currently only has newProductBatch form. Added a button on portalLanding for the productBatch template.

* No need for link

* Update pom.xml

no need for javax dependencies or version change

* Update application.yml

No need to change db name or username

* Created a CompleteProductBatchEntry class to be a middle man between Product and ProductBatch for the table. Made delete functionality work for the ProductBatchController, and the template for the productBatch.html. Also changed the return for the delete functionality for the Product and ProductBatch to redirect back to the main page.

* Added error checking if you want to put a product that already exists.
Modified the template for the portalLanding to display a different title.
Modified the login, productBatch, and the portalLanding to a better layout.
Added "back" button on both the product and productBatch page to take the
user back to the portalLanding page.
Modified the title of the product list on the product page to make it better.

* Modified StorageTransaction with some more attributes. Created a transaction
template that is now mapped to it's controller. Added two new .css files
for the loginForm and portalLanding for styling. Removed the inventory button
in the portalLanding.

* Made the functionality to update quantity/location of a particular product
batch from the batchID. Also added those options in the template of productBatch.
Also working on the template of a list of transactions to show a list of them.