Styling CSS merge, updates made by Nick Fritz (#7)
* Added a table for the transactions and added an attribute to the
transaction pojo for the type of transaction it is. New style sheet
that needs to be fixed. Removed some code that was not needed in
the controller.

* changed java version

* changing db login settings

* Added a table for the transactions and added an attribute to the
transaction pojo for the type of transaction it is. New style sheet
that needs to be fixed. Removed some code that was not needed in
the controller.

* Edited style.css to style products.html, productbatch.html, and transaction.html. Also did minor html changes. -Nick Fritz

* need to disregard change to db config

* Master branch runs on 1.8 java